-update to Xcode 6.1 (fix compile errors) -add cocos2d lib as a dependency in the app target

+++ cocos2d/libs/cocos2d/ccMacros.h
@@ -29,6 +29,7 @@
 
 #import <Foundation/Foundation.h>
 #import <Availability.h>
+#import <CoreGraphics/CGBase.h>
 
 /**
  @file
@@ -220,7 +221,7 @@ do {															\
  On Mac it returns 1;
  On iPhone it returns 2 if RetinaDisplay is On. Otherwise it returns 1
  */
-extern float __ccContentScaleFactor;
+extern CGFloat __ccContentScaleFactor;
 #define CC_CONTENT_SCALE_FACTOR() __ccContentScaleFactor
